requirements.txt
python-telegram-bot
telethon
requests
scikit-learn
tensorflow
gunicorn
